sistema_academico/requirements.txt
fastapi==0.75.1
uvicorn==0.17.6
sqlalchemy==1.4.36
pydantic==1.9.0
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.5
bcrypt==3.2.0
alembic==1.7.7
jinja2==3.1.1
aiofiles==0.8.0
python-dotenv==0.20.0
PyPDF2==2.6.0
python-docx==0.8.11
pillow==9.1.0
email-validator==1.3.0